refactor(tests): split tests
