+++ tests/grammar/test_grammar_func_call.py
@@ -1,0 +1,7 @@
+"Test func_call"
+
+from pytopas import grammar as g
+from tests.grammar.func_call_params import func_call_params
+from tests.helpers import make_trivial_grammar_test
+
+test_func_call = make_trivial_grammar_test(g.func_call, *func_call_params)
