+++ tests/grammar/test_grammar_integer.py
@@ -1,0 +1,6 @@
+"Test integer"
+
+from pytopas import grammar as g
+from tests.helpers import make_trivial_grammar_test
+
+test_interger = make_trivial_grammar_test(g.integer, ("100", None, {"integer": 100}))
