+++ tests/grammar/test_grammar_num_runs.py
@@ -1,0 +1,9 @@
+"Test local"
+from pytopas import ast
+from pytopas import grammar as g
+from tests.helpers import make_trivial_grammar_test
+
+test_num_runs = make_trivial_grammar_test(
+    g.num_runs,
+    ("num_runs 10", None, {"num_runs": ast.NumRunsNode(10)}),
+)
